organize some methods
